change temp location to os specific user directory

+++ main.py
@@ -6,15 +6,17 @@
 """
 import os
 import sys
+from appdirs import *
 from glob import glob
 from PySide6.QtCore import QPoint
 from BDRC.MVVM.view import AppView
 from BDRC.MVVM.model import OCRDataModel, SettingsModel
 from BDRC.MVVM.viewmodel import DataViewModel, SettingsViewModel
-from Config import TMP_DIR, read_settings
+from Config import read_settings
 from BDRC.Utils import get_screen_center, get_platform, create_dir
 from PySide6.QtWidgets import QApplication
 from BDRC.Styles import DARK
+
 
 
 if __name__ == "__main__":
@@ -22,7 +24,12 @@ if __name__ == "__main__":
     app = QApplication()
     app.setStyleSheet(DARK)
 
-    app_settings, ocr_settings = read_settings()
+    app_name = "OCR"
+    app_author = "BDRC"
+    user_data_dir = user_data_dir(app_name, app_author)
+    create_dir(user_data_dir)
+
+    app_settings, ocr_settings = read_settings(user_data_dir)
     print(f"starting with model_path: {app_settings.model_path}")
 
     data_model = OCRDataModel()
@@ -36,18 +43,23 @@ if __name__ == "__main__":
     app_view = AppView(
         dataview_model,
         settingsview_model,
-        platform
+        platform,
+        user_data_dir
     )
-
     app_view.resize(screen_data.start_width, screen_data.start_height)
     app_view.move(QPoint(screen_data.start_x, screen_data.start_y))
 
     # just delete tmp files on startup
-    create_dir(TMP_DIR)
-    tmp_files = glob(f"{TMP_DIR}/*")
+    tmp_img_dir = os.path.join(user_data_dir, "tmp", "images")
+    print(f"Tmp image dir: {tmp_img_dir}")
 
-    if len(tmp_files) > 0:
-        for file in tmp_files:
-            os.remove(file)
+    if os.path.isdir(tmp_img_dir):
+        tmp_files = glob(f"{tmp_img_dir}/*")
+        print(f"Tmp files: {len(tmp_files)}")
+        if len(tmp_files) > 0:
+            for file in tmp_files:
+                os.remove(file)
+    else:
+        print("tmp dir not a valid dir")
 
     sys.exit(app.exec())
